requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
better-profanity==0.7.0
black==21.7b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
demoji==1.1.0
discord-interactions==0.3.0
discord-py-slash-command==3.0.1
discord.py==1.7.3
dnspython==1.16.0
idna==3.2
importlib-resources==5.2.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
pycparser==2.20
Pygments==2.9.0
pymongo==3.12.0
PyNaCl==1.4.0
python-dotenv==0.19.0
regex==2021.8.3
requests==2.26.0
rich==10.7.0
six==1.16.0
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
zipp==3.5.0
